moved linear_system fixture from test_linalg.py to common fixture_setup.py
