+++ tests/test_linalg.py
@@ -7,22 +7,6 @@ import brahmap
 from brahmap.math.linalg import cg, parallel_norm
 
 
-@pytest.fixture
-def linear_system():
-    rng = np.random.default_rng(seed=867 + brahmap.MPI_UTILS.rank)
-    mat_size = 40
-    A = rng.random(size=(mat_size, mat_size))
-    A = A + A.T  # A symmetric matrix
-    A = A + mat_size * np.eye(mat_size)  # A positive definite matrix
-
-    # numpy array to sparse linop
-    A_op = brahmap.base.aslinearoperator(A)
-
-    # RHS vector
-    b = rng.random(mat_size)
-    return A_op, b
-
-
 def test_cg_against_scipy(linear_system):
     A_op, b = linear_system
